Add primitive Value Type reference projection support (part 1)
1. There are 2 ram classes generated from a primitive value type. One is
the L type ram class and the other is the Q type ram class. Add fields
Ltype and Qtype into J9Class. No non-primitive value classes, these 2
fields points to the J9Class itself.
2. For primitive VT, only the Q type can be instanciated. Do not
calculate the vtable, itable, ramMethods and ram constant pool of its
Ltype. The L and Q type share the same initialization status.
3. Add new flag J9_FINDCLASS_FLAG_QTYPE to indicate if it is a Q type
that is to be found or generated.
4. Both Q type and L type ram classes are added to the class table. A
new local type TYPE_QTYPE_RAM_CLASS in KeyHashTable.c is added. 
5. The same as existing classes, both L type and Q type ram classes will
be keyed on the class name in the class hash table. However, if it is a
Q type ram class, its hash value will be incremented by 1 (the value of
TYPE_QTYPE_RAM_CLASS)
6. Add J9_HASH_TABLE_QUERY_FLAG_QTYPE to indicate if the hash table
query is for a Q type or not.
7. Add J9CLASS_WALK_FLAG_SKIP_VT_LTYPE to indicate if the class segment
walk should skip the L type of primitive VT.

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>